// types.ts
export interface Experience {
  title: string;
  company: string;
  period: string;
  description: string[];
}

export interface Education {
  degree: string;
  institution: string;
  period: string;
}

export interface SkillCategory {
  name: string;
  skills: string[];
}

export interface Interest {
  name: string;
  icon: string;
}